requirements.txt
# ============================================
# Python Dependencies
# ============================================
# File: requirements.txt
# Install: pip install -r requirements.txt
# ============================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ML Model
xgboost==2.1.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1

# Database
supabase>=2.11.0

# Environment
python-dotenv>=1.0.1

# Data Validation
pydantic==2.10.4

# Database (Supabase)
supabase==2.11.0

# Environment Variables
python-dotenv==1.0.1
